prepare extraction of RemoteGit (rename ProjectTypeFilter to ComposerProjectFilter hide, IncludeIfHasFileFilter to RequiredFileFilter and hide GitlabNamespaceFilter)
